refactor transcribe-and-ingest-audio into several files; add second pinecone index name for ingestion
