+++ config/pinecone.ts
@@ -6,8 +6,13 @@ if (!process.env.PINECONE_INDEX_NAME) {
   throw new Error('Missing Pinecone index name in .env file');
 }
 
+if (!process.env.PINECONE_INGEST_INDEX_NAME) {
+  throw new Error('Missing Pinecone ingest index name in .env file');
+}
+
 const PINECONE_INDEX_NAME = process.env.PINECONE_INDEX_NAME ?? '';
+const PINECONE_INGEST_INDEX_NAME = process.env.PINECONE_INGEST_INDEX_NAME ?? '';
 
 // const PINECONE_NAME_SPACE = 'pdf-test'; //namespace is optional for your vectors
 
-export { PINECONE_INDEX_NAME };
+export { PINECONE_INDEX_NAME, PINECONE_INGEST_INDEX_NAME };
